requirements.txt
pyyaml
pymongo
pyautogui
numpy
pypiwin32
opencv-contrib-python
